+++ magda-web-client/craco.config.js
@@ -63,6 +63,11 @@ module.exports = {
                 })
             );
 
+            webpackConfig.externals = {
+                react: "React",
+                "react-dom": "ReactDOM"
+            };
+
             return webpackConfig;
         }
     }