+++ magda-web-client/src/polyfill.js
@@ -1,0 +1,8 @@
+// es6-shim is loaded from index.html
+//import "es6-shim";
+import "raf/polyfill";
+import "core-js/fn/symbol/iterator";
+import "core-js/es6/symbol";
+import "core-js/fn/object/entries";
+import "core-js/fn/object/values";
+import "./helpers/scroll-polyfill";